refactor: web adapter dto 패키지 분리
