+++ src/main/java/me/coonect/coonect/member/adapter/in/web/dto/request/EmailVerificationConfirmRequest.java
@@ -1,4 +1,4 @@
-package me.coonect.coonect.member.adapter.in.web;
+package me.coonect.coonect.member.adapter.in.web.dto.request;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
