+++ src/main/java/me/coonect/coonect/member/adapter/in/web/dto/response/MemberResponse.java
@@ -1,4 +1,4 @@
-package me.coonect.coonect.member.adapter.in.web;
+package me.coonect.coonect.member.adapter.in.web.dto.response;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
