+++ src/test/java/me/coonect/coonect/member/adapter/in/web/MemberSignupControllerTest.java
@@ -3,6 +3,8 @@ package me.coonect.coonect.member.adapter.in.web;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
+import me.coonect.coonect.member.adapter.in.web.dto.request.MemberSignupRequest;
+import me.coonect.coonect.member.adapter.in.web.dto.response.MemberResponse;
 import me.coonect.coonect.mock.TestContainer;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
